Test syslog collectors and move http handler to api
